// utils/types.ts
export type TechType = {
    image: string,
    name: string,
    size: string
}

export type Persona = {
    name: string;
    profession: string;
    details: string[];
    goals: string[];
    bio: string;
    frustrations: string;
    image: string;
}

export type EmpathyMap = {
    title: string;
    details: string[];
}